# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Logistics Microservices API
  version: 2.0.0
  description: |
    Complete endpoint inventory for Logistics Microservices
    (Auth, Order, Pricing, Tracking, Hub, Courier, Notification).

    ## Standard Response Format

    All API responses (except /health) follow a standard envelope:

    **Success:**
    ```json
    {
      "success": true,
      "data": { ... }
    }
    ```

    **Error:**
    ```json
    {
      "success": false,
      "error": {
        "code": "BAD_REQUEST",
        "message": "description of the error"
      }
    }
    ```
servers:
  - url: /
    description: Base URL for all microservices
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TrackingStatus:
      type: string
      enum:
        - CREATED
        - IN_HUB
        - IN_TRANSIT
        - OUT_DELIVERY
        - DELIVERED
        - FAILED
        - RETURNED
    ServiceType:
      type: string
      enum:
        - REGULAR
        - EXPRESS
        - SAMEDAY
        - NEXTDAY

    # ----- Standard Envelope Schemas -----
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: BAD_REQUEST
        message:
          type: string
          description: Human-readable error message
          example: invalid request body
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorDetail'

    # ----- Auth Schemas -----
...
    OrderRequest:
      type: object
      required: [sender_name, recipient_name, origin, destination, weight, dimensions, item_type, service_type, distance]
      properties:
        sender_name:
          type: string
        recipient_name:
          type: string
        origin:
          type: string
        destination:
          type: string
        weight:
          type: number
        dimensions:
          type: string
        item_type:
          type: string
        service_type:
          $ref: '#/components/schemas/ServiceType'
        distance:
          type: number
    OrderResponseData:
      type: object
      properties:
        order_id:
          type: string
        resi_id:
          type: string
        status:
          $ref: '#/components/schemas/TrackingStatus'
        total_cost:
          type: number
    OrderResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/OrderResponseData'
    ValidateResiResponseData:
      type: object
      properties:
        status:
          type: string
          example: valid
    ValidateResiResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/ValidateResiResponseData'

    # ----- Pricing Schemas -----
...
paths:
...
  /order:
    post:
      summary: Create a new order
      tags: [Order]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{resiID}/validate:
    get:
      summary: Validate Resi ID
      tags: [Order]
      parameters:
        - in: path
          name: resiID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resi ID is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResiResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resi ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # -------------------------
  # Pricing Service
  # -------------------------
...